+++ PADL/src/main/java/padl/util/adapter/InterfaceAdapter.java
@@ -11,18 +11,9 @@
 package padl.util.adapter;
 
 import java.util.Iterator;
-import padl.event.IEvent;
-import padl.event.IModelListener;
-import padl.kernel.FirstClassAdapter;
-import padl.kernel.IConstituent;
-import padl.kernel.IConstituentExtension;
-import padl.kernel.IConstituentOfEntity;
-import padl.kernel.IElement;
-import padl.kernel.IFilter;
 import padl.kernel.IFirstClassEntity;
 import padl.kernel.IInterface;
 import padl.kernel.impl.Factory;
-import padl.visitor.IVisitor;
 
 /**
  * @author Yann-Gaël Guéhéneuc
